Add new plugin to detect Nginx version disclosure and enhance add_header_redefinition
Introduced a new plugin that checks if the Nginx version is being disclosed which is a security vulnerability. At the same time, the previous "add_header_redefinition" plugin has been enhanced to check if any secure headers were dropped. Severity level of this issue is now determined based on whether a secure header was dropped or not.